Split up Boogie AST
