+++ libsolidity/CMakeLists.txt
@@ -55,6 +55,9 @@ set(sources
 	boogie/ASTBoogieStats.cpp
 	boogie/ASTBoogieUtils.cpp
 	boogie/BoogieAst.cpp
+	boogie/BoogieAstDecl.cpp
+	boogie/BoogieAstExpr.cpp
+	boogie/BoogieAstStmt.cpp
 	boogie/BoogieContext.cpp
 	boogie/EmitsChecker.cpp
 	boogie/StoragePtrHelper.cpp
