+++ libsolidity/boogie/ASTBoogieUtils.cpp
@@ -181,10 +181,10 @@ bg::ProcDeclRef ASTBoogieUtils::createCallProc(BoogieContext& context)
 	thenBlock->addStmt(bg::Stmt::assign(
 			context.boogieBalance()->getRefTo(),
 			bg::Expr::arrupd(context.boogieBalance()->getRefTo(), context.boogieThis()->getRefTo(), addBalance.expr)));
-	thenBlock->addStmt(bg::Stmt::assign(result, bg::Expr::lit(true)));
+	thenBlock->addStmt(bg::Stmt::assign(result, bg::Expr::true_()));
 	// Unsuccessful transfer
 	bg::Block::Ref elseBlock = bg::Block::block();
-	elseBlock->addStmt(bg::Stmt::assign(result, bg::Expr::lit(false)));
+	elseBlock->addStmt(bg::Stmt::assign(result, bg::Expr::false_()));
 	// Nondeterministic choice between success and failure
 	bg::Block::Ref callBlock = bg::Block::block();
 	callBlock->addStmt(bg::Stmt::comment("TODO: call fallback"));
@@ -258,11 +258,11 @@ bg::ProcDeclRef ASTBoogieUtils::createSendProc(BoogieContext& context)
 	thenBlock->addStmt(bg::Stmt::assign(
 			context.boogieBalance()->getRefTo(),
 			bg::Expr::arrupd(context.boogieBalance()->getRefTo(), context.boogieMsgSender()->getRefTo(), subSenderBalance.expr)));
-	thenBlock->addStmt(bg::Stmt::assign(result, bg::Expr::lit(true)));
+	thenBlock->addStmt(bg::Stmt::assign(result, bg::Expr::true_()));
 
 	// Unsuccessful transfer
 	bg::Block::Ref elseBlock = bg::Block::block();
-	elseBlock->addStmt(bg::Stmt::assign(result, bg::Expr::lit(false)));
+	elseBlock->addStmt(bg::Stmt::assign(result, bg::Expr::false_()));
 
 	bg::Block::Ref transferBlock = bg::Block::block();
 	// Precondition: there is enough ether to transfer
@@ -351,7 +351,7 @@ ASTBoogieUtils::ExprWithCC ASTBoogieUtils::encodeArithBinaryOp(BoogieContext& co
 			auto lhsLit = dynamic_pointer_cast<bg::IntLit const>(lhs);
 			auto rhsLit = dynamic_pointer_cast<bg::IntLit const>(rhs);
 			if (lhsLit && rhsLit)
-				result = bg::Expr::lit(boost::multiprecision::pow(lhsLit->getVal(), rhsLit->getVal().convert_to<unsigned>()));
+				result = bg::Expr::intlit(boost::multiprecision::pow(lhsLit->getVal(), rhsLit->getVal().convert_to<unsigned>()));
 			else
 			{
 				context.reportError(associatedNode, "Exponentiation is not supported in 'int' encoding");
@@ -457,9 +457,9 @@ ASTBoogieUtils::ExprWithCC ASTBoogieUtils::encodeArithBinaryOp(BoogieContext& co
 	}
 	case BoogieContext::Encoding::MOD:
 	{
-		auto modulo = bg::Expr::lit(boost::multiprecision::pow(bg::bigint(2), bits));
-		auto largestSigned = bg::Expr::lit(boost::multiprecision::pow(bg::bigint(2), bits - 1) - 1);
-		auto smallestSigned = bg::Expr::lit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
+		auto modulo = bg::Expr::intlit(boost::multiprecision::pow(bg::bigint(2), bits));
+		auto largestSigned = bg::Expr::intlit(boost::multiprecision::pow(bg::bigint(2), bits - 1) - 1);
+		auto smallestSigned = bg::Expr::intlit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
 		switch (op)
 		{
 		case Token::Add:
@@ -494,8 +494,8 @@ ASTBoogieUtils::ExprWithCC ASTBoogieUtils::encodeArithBinaryOp(BoogieContext& co
 			auto prod = bg::Expr::times(lhs, rhs);
 			if (isSigned)
 			{
-				auto lhs1 = bg::Expr::cond(bg::Expr::gte(lhs, bg::Expr::lit((long)0)), lhs, bg::Expr::plus(modulo, lhs));
-				auto rhs1 = bg::Expr::cond(bg::Expr::gte(rhs, bg::Expr::lit((long)0)), rhs, bg::Expr::plus(modulo, rhs));
+				auto lhs1 = bg::Expr::cond(bg::Expr::gte(lhs, bg::Expr::intlit((long)0)), lhs, bg::Expr::plus(modulo, lhs));
+				auto rhs1 = bg::Expr::cond(bg::Expr::gte(rhs, bg::Expr::intlit((long)0)), rhs, bg::Expr::plus(modulo, rhs));
 				auto prod = bg::Expr::mod(bg::Expr::times(lhs1, rhs1), modulo);
 				result = bg::Expr::cond(bg::Expr::gt(prod, largestSigned), bg::Expr::minus(prod, modulo), prod);
 			}
@@ -605,16 +605,16 @@ ASTBoogieUtils::ExprWithCC ASTBoogieUtils::encodeArithUnaryOp(BoogieContext& con
 			auto sub = bg::Expr::neg(subExpr);
 			if (isSigned)
 			{
-				auto smallestSigned = bg::Expr::lit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
+				auto smallestSigned = bg::Expr::intlit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
 				result = bg::Expr::cond(bg::Expr::eq(subExpr, smallestSigned),
 						smallestSigned,
 						sub);
 			}
 			else
 			{
-				auto modulo = bg::Expr::lit(boost::multiprecision::pow(bg::bigint(2), bits));
-				result = bg::Expr::cond(bg::Expr::eq(subExpr, bg::Expr::lit((long)0)),
-						bg::Expr::lit((long)0),
+				auto modulo = bg::Expr::intlit(boost::multiprecision::pow(bg::bigint(2), bits));
+				result = bg::Expr::cond(bg::Expr::eq(subExpr, bg::Expr::intlit((long)0)),
+						bg::Expr::intlit((long)0),
 						bg::Expr::minus(modulo, subExpr));
 			}
 			ecc = bg::Expr::eq(sub, result);
@@ -690,9 +690,9 @@ bg::Expr::Ref ASTBoogieUtils::checkImplicitBvConversion(bg::Expr::Ref expr, Type
 		if (auto exprLit = dynamic_pointer_cast<bg::IntLit const>(expr))
 		{
 			if (exprLit->getVal() < 0) // Negative literals are tricky
-				return context.bvNeg(targetBits, bg::Expr::lit(-exprLit->getVal(), targetBits));
+				return context.bvNeg(targetBits, bg::Expr::bvlit(-exprLit->getVal(), targetBits));
 			else
-				return bg::Expr::lit(exprLit->getVal(), targetBits);
+				return bg::Expr::bvlit(exprLit->getVal(), targetBits);
 		}
 		else if (isBitPreciseType(exprType))
 		{
@@ -778,8 +778,8 @@ bg::Expr::Ref ASTBoogieUtils::getTCCforExpr(bg::Expr::Ref expr, TypePointer tp, 
 	{
 		auto enumTp = dynamic_cast<EnumType const *>(tp);
 		solAssert(enumTp, "");
-		lowerBound = bg::Expr::lte(bg::Expr::lit((long)0), expr);
-		upperBound = bg::Expr::lt(expr, bg::Expr::lit(enumTp->enumDefinition().members().size()));
+		lowerBound = bg::Expr::lte(bg::Expr::intlit((long)0), expr);
+		upperBound = bg::Expr::lt(expr, bg::Expr::intlit(enumTp->enumDefinition().members().size()));
 	}
 	// Otherwise get smallest and largest
 	else if (isBitPreciseType(tp))
@@ -787,15 +787,15 @@ bg::Expr::Ref ASTBoogieUtils::getTCCforExpr(bg::Expr::Ref expr, TypePointer tp, 
 		unsigned bits = getBits(tp);
 		if (isSigned(tp))
 		{
-			auto largestSigned = bg::Expr::lit(boost::multiprecision::pow(bg::bigint(2), bits - 1) - 1);
-			auto smallestSigned = bg::Expr::lit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
+			auto largestSigned = bg::Expr::intlit(boost::multiprecision::pow(bg::bigint(2), bits - 1) - 1);
+			auto smallestSigned = bg::Expr::intlit(-boost::multiprecision::pow(bg::bigint(2), bits - 1));
 			lowerBound = bg::Expr::lte(smallestSigned, expr);
 			upperBound = bg::Expr::lte(expr, largestSigned);
 		}
 		else
 		{
-			auto largestUnsigned = bg::Expr::lit(boost::multiprecision::pow(bg::bigint(2), bits) - 1);
-			auto smallestUnsigned = bg::Expr::lit(long(0));
+			auto largestUnsigned = bg::Expr::intlit(boost::multiprecision::pow(bg::bigint(2), bits) - 1);
+			auto smallestUnsigned = bg::Expr::intlit(long(0));
 			lowerBound = bg::Expr::lte(smallestUnsigned, expr);
 			upperBound = bg::Expr::lte(expr, largestUnsigned);
 		}
@@ -812,7 +812,7 @@ bg::Expr::Ref ASTBoogieUtils::getTCCforExpr(bg::Expr::Ref expr, TypePointer tp, 
 	default:
 		solAssert(false, "");
 	}
-	return bg::Expr::lit(true);
+	return bg::Expr::true_();
 }
 
 bg::Expr::Ref ASTBoogieUtils::defaultValue(TypePointer type, BoogieContext& context)
@@ -857,7 +857,7 @@ ASTBoogieUtils::Value ASTBoogieUtils::defaultValueInternal(TypePointer type, Boo
 	{
 		// 0
 		auto lit = context.intLit(0, ASTBoogieUtils::getBits(type));
-		return {lit->toSMT2(), lit};
+		return {lit->toSMT2String(), lit};
 	}
 	case Type::Category::Address:
 	case Type::Category::Contract:
@@ -865,19 +865,19 @@ ASTBoogieUtils::Value ASTBoogieUtils::defaultValueInternal(TypePointer type, Boo
 	{
 		// 0
 		auto lit = context.intLit(0, 256);
-		return {lit->toSMT2(), lit};
+		return {lit->toSMT2String(), lit};
 	}
 	case Type::Category::Bool:
 	{
 		// False
-		auto lit = bg::Expr::lit(false);
-		return {lit->toSMT2(), lit};
+		auto lit = bg::Expr::false_();
+		return {lit->toSMT2String(), lit};
 	}
 	case Type::Category::FixedBytes:
 	{
 		auto fbType = dynamic_cast<FixedBytesType const*>(type);
 		auto lit = context.intLit(0, fbType->numBytes() * 8);
-		return {lit->toSMT2(), lit};
+		return {lit->toSMT2String(), lit};
 	}
 	case Type::Category::Struct:
 	{
@@ -928,7 +928,7 @@ ASTBoogieUtils::Value ASTBoogieUtils::defaultValueInternal(TypePointer type, Boo
 			args.push_back(arrLength);
 			auto bgExpr = bg::Expr::fn(arrConstr->getName(), args);
 			// SMT expression is constructed afterwards (not to be included in the const array function)
-			smtExpr = "(|" + arrConstr->getName() + "| " + smtExpr + " " + arrLength->toString() + ")";
+			smtExpr = "(|" + arrConstr->getName() + "| " + smtExpr + " " + arrLength->toBgString() + ")";
 			return {smtExpr, bgExpr};
 		}
 		break;
