+++ libsolidity/boogie/ASTBoogieUtils.h
@@ -2,6 +2,9 @@
 
 #include <libsolidity/ast/Types.h>
 #include <libsolidity/boogie/BoogieAst.h>
+#include <libsolidity/boogie/BoogieAstDecl.h>
+#include <libsolidity/boogie/BoogieAstExpr.h>
+#include <libsolidity/boogie/BoogieAstStmt.h>
 #include <liblangutil/Scanner.h>
 #include <string>
